+++ src/helpers/helpers.js
@@ -1,8 +1,6 @@
-const helpers = {
+module.exports = {
     // This helper is an example and has to be deleted in distribution
     myFirstHelper: function (options) {
         return options.fn(this);
     },
 }
-
-module.exports = helpers;