+++ src/jsx/actions/index.jsx
@@ -1,5 +1,0 @@
-export function myFirstAction() {
-    return (dispatch, getState) => {
-        dispatch({type: 'MY_FIRST_ACTION'});
-    }
-}