+++ src/jsx/reducers/index.jsx
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import myFirstReducer from './myFirstReducer';
-
-export default combineReducers({
-    myFirstReducer,
-});