+++ tasks/node-env.js
@@ -1,0 +1,6 @@
+// This NODEENV function:
+// Sets NODE_ENV = 'production'
+export default function nodeEnv(done) {
+    process.env.NODE_ENV = 'production';
+    done();
+}